+++ src/main/java/com/delfood/controller/ShopController.java
@@ -200,74 +200,7 @@ public class ShopController {
         HttpStatus.OK);
   }
 
-
-  /**
-   * 매장의 배달 가능 지역을 추가한다. 배달 가능 지역에 포함되어 있는 사용자에게 검색이 된다.
-   * 
-   * @author jun
-   * @param deliveryLocationInfo 새로운 배달지역 정보
-   * @param session 사용자의 세션
-   * @return
-   */
-  @PostMapping("deliveryLocation")
-  public ResponseEntity<AddDeliveryLocationResponse> addDeliveryLocation(
-      @RequestBody(required = true) AddDeliveryLocationRequest deliveryLocationInfo,
-      HttpSession session) {
-    String ownerId = SessionUtil.getLoginOwnerId(session);
-    Long shopId = deliveryLocationInfo.getShopId();
-    String townCode = deliveryLocationInfo.getTownCode();
-
-    // 로그인 하지 않았을시
-    if (ownerId == null) {
-      return new ResponseEntity<ShopController.AddDeliveryLocationResponse>(
-          AddDeliveryLocationResponse.NO_LOGIN, HttpStatus.UNAUTHORIZED);
-    }
-    // 로그인한 사장님이 해당 가게의 사장님이 아닐 시
-    if (shopService.isShopOwner(shopId, ownerId) == false) {
-      return new ResponseEntity<ShopController.AddDeliveryLocationResponse>(
-          AddDeliveryLocationResponse.UNAUTHORIZED, HttpStatus.UNAUTHORIZED);
-    }
-
-    // 핵심 로직
-    shopService.addDeliveryLocation(shopId, townCode);
-
-
-    return new ResponseEntity<ShopController.AddDeliveryLocationResponse>(
-        AddDeliveryLocationResponse.SUCCESS, HttpStatus.CREATED);
-  }
-
-  /**
-   * 배달 지역 삭제.
-   * 
-   * @author jun
-   * @param id 삭제할 배달 지역 id
-   * @param session 접속한 사용자의 세션
-   * @return
-   */
-  @DeleteMapping("deliveryLocation/{id}")
-  public ResponseEntity<DeleteDeliveryLocationResponse> deleteDeliveryLocation(
-      @PathVariable(required = true, value = "id") Long id, HttpSession session) {
-
-    String ownerId = SessionUtil.getLoginOwnerId(session);
-    // 로그인 하지 않았을시
-    if (ownerId == null) {
-      return new ResponseEntity<ShopController.DeleteDeliveryLocationResponse>(
-          DeleteDeliveryLocationResponse.NO_LOGIN, HttpStatus.UNAUTHORIZED);
-    }
-
-    // 자신의 가게가 아닐 시
-    if (shopService.isShopOwnerByDeliveryLocationId(id, ownerId) == false) {
-      return new ResponseEntity<ShopController.DeleteDeliveryLocationResponse>(
-          DeleteDeliveryLocationResponse.UNAUTHORIZED, HttpStatus.UNAUTHORIZED);
-    }
-
-    // 핵심 로직
-    shopService.deleteDeliveryLocation(id);
-
-
-    return new ResponseEntity<ShopController.DeleteDeliveryLocationResponse>(
-        DeleteDeliveryLocationResponse.SUCCESS, HttpStatus.OK);
-  }
+  
 
   /**
    * 매장 정보를 조회한다.
@@ -391,41 +324,9 @@ public class ShopController {
   }
 
 
-  @Getter
-  @RequiredArgsConstructor
-  private static class AddDeliveryLocationResponse {
-    enum Result {
-      SUCCESS, NO_LOGIN, UNAUTHORIZED
-    }
-
-    @NonNull
-    private Result result;
-    private static final AddDeliveryLocationResponse SUCCESS =
-        new AddDeliveryLocationResponse(Result.SUCCESS);
-    private static final AddDeliveryLocationResponse NO_LOGIN =
-        new AddDeliveryLocationResponse(Result.NO_LOGIN);
-    private static final AddDeliveryLocationResponse UNAUTHORIZED =
-        new AddDeliveryLocationResponse(Result.UNAUTHORIZED);
 
 
-  }
 
-  @Getter
-  @RequiredArgsConstructor
-  private static class DeleteDeliveryLocationResponse {
-    enum Result {
-      SUCCESS, NO_LOGIN, UNAUTHORIZED
-    }
-
-    @NonNull
-    private Result result;
-    private static final DeleteDeliveryLocationResponse SUCCESS =
-        new DeleteDeliveryLocationResponse(Result.SUCCESS);
-    private static final DeleteDeliveryLocationResponse NO_LOGIN =
-        new DeleteDeliveryLocationResponse(Result.NO_LOGIN);
-    private static final DeleteDeliveryLocationResponse UNAUTHORIZED =
-        new DeleteDeliveryLocationResponse(Result.UNAUTHORIZED);
-  }
 
   @Getter
   @RequiredArgsConstructor
@@ -452,14 +353,7 @@ public class ShopController {
     private Long lastId;
   }
 
-  @Getter
-  @Setter
-  private static class AddDeliveryLocationRequest {
-    @NonNull
-    private Long shopId;
-    @NonNull
-    private String townCode;
-  }
+  
 
 
 }
